Refactor validations
