+++ src/test-run/commands/prop-validations/factories.js
@@ -1,0 +1,40 @@
+// -------------------------------------------------------------
+// WARNING: this file is used by both the client and the server.
+// Do not use any browser or node-specific API!
+// -------------------------------------------------------------
+
+export function createIntegerValidator (ErrorCtor) {
+    return (name, val) => {
+        var valType = typeof val;
+
+        if (valType !== 'number')
+            throw new ErrorCtor(name, valType);
+
+        var isInteger = !isNaN(val) &&
+                        isFinite(val) &&
+                        val === Math.floor(val);
+
+        if (!isInteger)
+            throw new ErrorCtor(name, val);
+    };
+}
+
+export function createPositiveIntegerValidator (ErrorCtor) {
+    var integerValidator = createIntegerValidator(ErrorCtor);
+
+    return (name, val) => {
+        integerValidator(name, val, ErrorCtor);
+
+        if (val < 0)
+            throw new ErrorCtor(name, val);
+    };
+}
+
+export function createBooleanValidator (ErrorCtor) {
+    return (name, val) => {
+        var valType = typeof val;
+
+        if (valType !== 'boolean')
+            throw new ErrorCtor(name, valType);
+    };
+}
