--- test/server/data/expected-test-run-errors/action-integer-option-error
@@ -1,19 +1,0 @@
-Chrome 15.0.874 / Mac OS X 10.8.1
-The offsetX option is expected to be an integer, but it was NaN.
-
-   18 |function func1 () {
-   19 |    record = createCallsiteRecord('func1');
-   20 |}
-   21 |
-   22 |(function func2 () {
- > 23 |    func1();
-   24 |})();
-   25 |
-   26 |stackTrace.filter.deattach(stackFilter);
-   27 |
-   28 |module.exports = record;
-
-   at func2 (testfile.js:23:5)
-   at Object.<anonymous> (testfile.js:24:3)
-
-Screenshot: /unix/path/with/<tag>
